pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "probe-benchmark"
version = "0.1.0"
description = "PROBE: Proactive Resolution of Bottlenecks Benchmark"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "datasets>=2.14.0",
    "huggingface-hub>=0.17.0",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.28.0",
    "numpy>=1.24.0",
]

baselines = [
    "litellm>=1.0.0",
    "numpy>=1.24.0",
]

all = [
    "streamlit>=1.28.0",
    "numpy>=1.24.0",
    "litellm>=1.0.0",
]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = ["-W", "ignore"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["data_generation", "configs", "evaluation", "annotation", "baselines"]

[tool.hatch.build.targets.wheel]
packages = ["."]
